# Optimize some methods in RightsManagementEncryptionTransform, reduce allocs