## Description

Optimizes some low-hanging fruit in this mostly unused part of WPF to remove some unnecessary allocations.

- `CompareOrdinal` replaced with the most obvious `StartsWith`.
- The Base32 wannabe can be now encoded into a stackallocated buffer if its big enough.
- Parsing authentication type now does it directly off the sliced input string.
- Removed the `LicenseStreamNamePrefixLength`, 4 bytes saved sounds good.

Since I like numbers, here's a benchmark of `MakeUseLicenseStreamName`.

| Method    |Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|---------- |-----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original  |   565.6 ns |    1.71 ns |     1.43 ns | 0.0172 |         260 B |         288 B |
| PR__EDIT2 |  541.0 ns |    2.65 ns |     2.07 ns | 0.0048 |         386 B |          88 B |

## Customer Impact

Decreased allocations, improved performance in this edge-case such as anyone still using this.

## Regression

No.

## Testing

Local build, didn't test the full functionality itself but at least `MakeUseLicenseStreamName` + `Base32EncodeWithoutPadding` outputs have been tested to provide same set of values before and after. However the swaps are pretty much 1:1.

## Risk

Low.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9851)